backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
motor==3.7.0
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.20
pyjwt==2.10.1
passlib==1.7.4
bcrypt==4.2.1
